Lots

+++ StreamMasterInfrastructure.EF/Repositories/VideoStreamLinkRepository.cs
@@ -55,7 +55,7 @@ public class VideoStreamLinkRepository : RepositoryBase<VideoStreamLink>, IVideo
     public VideoStreamLink GetVideoStreamLink(string ParentVideoStreamId, string ChildVideoStreamId, int? Rank)
     {
         Rank ??= Count();
-        return new VideoStreamLink { ParentVideoStreamId = ParentVideoStreamId, ChildVideoStreamId = ChildVideoStreamId, Rank = Rank };
+        return new VideoStreamLink { ParentVideoStreamId = ParentVideoStreamId, ChildVideoStreamId = ChildVideoStreamId, Rank = (int)Rank };
     }
 
     public async Task AddVideoStreamTodVideoStream(string ParentVideoStreamId, string ChildVideoStreamId, int? Rank, CancellationToken cancellationToken)
